+++ src/web/extension.ts
@@ -1,0 +1,9 @@
+import * as vscode from "vscode";
+import { activateTableTestExtension } from "../activateExtension";
+import { WebJavaFormatterProfileReader } from "../webJavaFormatterProfileReader";
+
+export function activate(context: vscode.ExtensionContext) {
+  activateTableTestExtension(context, new WebJavaFormatterProfileReader());
+}
+
+export function deactivate() {}